+++ repository/api/src/main/java/com/expediagroup/streamplatform/streamregistry/repository/ConsumerRepository.java
@@ -20,5 +20,4 @@ import com.expediagroup.streamplatform.streamregistry.model.keys.ConsumerKey;
 import com.expediagroup.streamplatform.streamregistry.model.keys.StreamKey;
 
 public interface ConsumerRepository extends Repository<Consumer, ConsumerKey> {
-  void findAllAndDelete(StreamKey key);
 }