+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Implemented entity delete functionality.
+
 ## [0.17.0] - 2021-01-29
 ### Changed
 - Enhanced `EntityView` to support the retrieval of deleted `Entity`s